Fixed a bug concerning reading raw files -- it _really_ read raw files, returning both a buffer with the contents and the buffer size.
Now it returns the null-terminated buffer, which is to be
interpreted as a C string.  Note that this will produce undesired
results if files already contain null characters: the Tmake routine
will "truncate" them when constructing a corresponding term.

There is no trivial way of fixing this: Tmake (unlike the old
TBmake) does not have a "binary string" placeholder: to
generalize, here, Tmake should be extended to allow arbitrary
binary buffers.