The reject mechanism was `optimizing' rather too optimistically: it threw away "ambiguity clusters" when resolving them into non-ambiguity.
However, since the `wrong' stack can be the one to live on in this case,
this may result in reject-disambiguation turning out to `pick out the
wrong one': instead of looking up the right replacement, the (rejected)
term could be reused.

The fix consists of keeping full ambiguity information at all times,
and allowing an ambiguity lookup to return a list of any non-zero
length.