Ambiguities are introduced and counted when parsing.  They can disappear erratically, if a stack containing one has no continuation, or by a reject production.  Finally, by explicit `supersort restriction' a top-level ambiguity might get resolved.
The ambiguity count after parsing is an upper limit for that
reason.  However, some tweaks have been implemented that allow
returning an exact count at all times, except when no output
is requested.  In that case, the (essential) forest yielding traversal
isn't performed, which rules out the possibility of determining
the exact number of ambiguities.