This version contains a scheme for Multiset Priorisation not under control.
A multiset is a function Ms: prod --> occurrence.
Multisets are constructed from trees by counting productions in them.

Multiset ordering (`>>>') is defined on trees/multisets as
follows, using some local priority ordering `>>':

Trees:

	t1 >>> t2	<==>	MS(t1) >>> MS(t2)

Multisets:

	M >>> N	<==>
	        M != N
	and
	        for all y in M
		        if M(y) > N(y)
			        there exists a x in N s.t.
				        M(x) < N(x)
			        and
				        ! ( x >> y)

This is a digression from the ordering as is described, which is:

	M >>> N    <==>
		M != N
	and
	        for all y in M
		        if M(y) > N(y)
			        there exists a x in N s.t.
				        M(x) < N(x)
			        and
				        y >> x

A lot more investigation into this particular form of priority
disambiguation is required.