The (recent) GNU C library contains a `strndup' function:
	char *strndup(const char *s, size_t size)

This collides fatally with the function

	char *strndup(char *bgn, char *end)

as it existed in toolbus' source.

The two functions are functionally identical, so strndup has been
redesigned to conform to its API.

After doing this (of course) I noted that there was actually only a single
call to `strndup', which, because it was always called with exactly the
length of the string to be duplicated as an argument, could just as
well have been replaced by a plain `strdup'.  Sigh.