The genmakefile.sh  script wrote Makefile and  Makefile.in from template that was entirely in-line in the sh script.
This in-line  template contained  sh- and  make variables  (that, hence,
needed extra escaping).  It  contained configure variables (that, hence,
needed to be protected from getting substituted away by configure).

Adding to this misery, linux has no truly Bourne-compatible sh; on linux
systems, sh happens to be bash.  And  bash is a shell on which $(str) is
equivalent to `str`: command substitution.

To get this  working at all for  both bash and old-fashioned  sh, and to
unmuddle the multi-level  escaping somewhat, the in-line  template is no
more.  Instead,  a separate template file  is used, so that  any and all
interference  with implicit  substitution  mechanisms  in the  scripting
language used is avoided.