Robuster dependency determination.  Ick.  Ick.  Ick.
Sometimes the Gnu C Compiler is available, sometimes it isn't.
Sometimes it's called `gcc', sometimes it isn't.  Some systems have the
(X11R4) tool `makedepend', some don't. On some systems that have
`makedepend', it is consistent with the C development stuff's setup, on
others it isn't.  And on some systems there's a `gccmakedep', sometimes
replacing `makedepend', sometimes not.

Whenever a Gnu C Compiler (be it in disguise or not) is detected, it's
used to generate dependencies.  Otherwise, dependency generation falls
back on `gccmakedep', `makedepend', in that order.

If somebody wishes to write an Autoconfig macro to test whether the
current non-Gnu compiler can generate dependencies, well, be my guest.