SDF syntax change.  Consider the production:
	"%" ( "a" | "b" ) -> S

This was parsed as implicit prefix function notation instead of the
intended %-literal followed by either a S1 or a S2.  As a rather
unexpected result, the input terms "%a" and "%b" do not parse,
but "%(a)" and "%(b)" do.  Ewk.

Confusingly, a nearly identical way of defining the intended syntax

	[\%] ( S1 | S2 ) -> S

produced the expected parses.  Ewk, again.

Quoted literals are now disallowed in "function prefix" notation.  Do
write them unquoted.  Or else.