Resolved bug #86 The semantics of  comprehensions has been extended as follows:
(a) The syntax of comprehensions has been extended to allow more than one
    expression, i.e. {E1, E2, ..., En | Gen1, ... Genm }

(b) During each iteration the values of E1, ... are combined (using the union
    operator) with the bag accumulated so far.

This solution excludes some expressions (e.g. the case that a *variable*
number of values is added per iteration) but this cleaner sementics is
preferable.