A major remake of the rviz tool:
(a) Selecting a relation in the relation pane now enables/disables
    entries in the itemtypes pane (= the itemtypes that may appear in the
    selected relation)
(b) Selecting an enabled itemtype fills the list of items of the desired itemtype.

This gives a natural way of playing with the relational info. The highlighting
of list items still needs polishing.

The file view and tree map view are now integrated in the same user interface.
A button allows switching between the two (synchronized) views.